+++ heat/tests/test_get_grid_y.f90
@@ -8,10 +8,10 @@ program test_get_grid_y
 
   integer, parameter :: grid_id = 1
   integer, parameter :: ny = 1
-  real, parameter, dimension(ny) :: expected_y = (/ 0.0 /)
+  double precision, parameter, dimension(ny) :: expected_y = (/ 0.0 /)
 
   type (bmi_heat) :: m
-  real, dimension(ny) :: grid_y
+  double precision, dimension(ny) :: grid_y
   integer :: i
 
   status = m%initialize(config_file)
